utilities/legacy/snapshot/requirements.txt
typer == 0.15.1
python-bitcoinlib == 0.12.2


# Format and lint
black == 22.3
ruff == 0.9.7